+++ app/src/test/java/dev/havoc/taxihud/phone/TaxiNotificationListenerServiceTest.java
@@ -4,6 +4,7 @@ import static org.junit.Assert.assertEquals;
 
 import android.app.Notification;
 import android.os.Process;
+import android.os.UserHandle;
 import android.service.notification.StatusBarNotification;
 
 import org.junit.Test;
@@ -42,7 +43,34 @@ public final class TaxiNotificationListenerServiceTest {
                                 RuntimeEnvironment.getApplication())).size());
     }
 
+    @Test
+    public void packageChoiceAppliesBeforeReadingNotificationsFromEveryDeliveredProfile() {
+        dev.havoc.taxihud.phone.config.AdapterRepository repository =
+                new dev.havoc.taxihud.phone.config.AdapterRepository(
+                        RuntimeEnvironment.getApplication());
+        repository.setPackageAllowed("ru.yandex.go", false);
+
+        StatusBarNotification personal = notification(
+                "ru.yandex.go", 1, 1_000L, Process.myUserHandle());
+        StatusBarNotification work = notification(
+                "ru.yandex.go", 2, 2_000L, UserHandle.getUserHandleForUid(1_010_000));
+        StatusBarNotification privateProfile = notification(
+                "ru.yandex.go", 3, 3_000L, UserHandle.getUserHandleForUid(1_110_000));
+
+        assertEquals(0, TaxiNotificationListenerService.selectActiveConfiguredNotifications(
+                new StatusBarNotification[] {personal, work, privateProfile}, repository).size());
+
+        repository.setPackageAllowed("ru.yandex.go", true);
+        assertEquals(3, TaxiNotificationListenerService.selectActiveConfiguredNotifications(
+                new StatusBarNotification[] {personal, work, privateProfile}, repository).size());
+    }
+
     private static StatusBarNotification notification(String packageName, int id, long postTime) {
+        return notification(packageName, id, postTime, Process.myUserHandle());
+    }
+
+    private static StatusBarNotification notification(
+            String packageName, int id, long postTime, UserHandle user) {
         return new StatusBarNotification(
                 packageName,
                 packageName,
@@ -52,7 +80,7 @@ public final class TaxiNotificationListenerServiceTest {
                 20_000,
                 0,
                 new Notification(),
-                Process.myUserHandle(),
+                user,
                 postTime);
     }
 }